refactoring the tests to make them part of the existing test and reduce noise
