+++ backend/tests/app/data_products/test_router.py
@@ -2,6 +2,7 @@ import json
 import uuid
 from copy import deepcopy
 from typing import Any, Dict
+from unittest.mock import AsyncMock, patch
 from uuid import UUID
 
 import pytest
@@ -11,6 +12,7 @@ from app.authorization.roles.schema import Scope
 from app.core.authz import Action
 from app.resource_names.service import ResourceNameValidityType
 from app.settings import settings
+from tests.app.core.webhooks.helpers import webhook_v2_config
 from tests.factories import (
     DataOutputDatasetAssociationFactory,
     DataProductFactory,
@@ -773,3 +775,162 @@ class TestDataProductsRouter:
     @staticmethod
     def get_data_product_settings(client: TestClient, data_product_id: UUID):
         return client.get(f"{ENDPOINT}/{data_product_id}/settings")
+
+
+class TestDataProductV2Events:
+    invalid_id = "00000000-0000-0000-0000-000000000000"
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_created_fires_event(self, mock_webhook, client, session, payload):
+        mock_webhook.return_value = AsyncMock()
+        RoleService(db=session).initialize_prototype_roles()
+
+        with webhook_v2_config():
+            response = client.post(ENDPOINT, json=payload)
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.created"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_updated_fires_event(self, mock_webhook, client, payload):
+        mock_webhook.return_value = AsyncMock()
+        dp = DataProductFactory()
+
+        with webhook_v2_config():
+            response = client.put(f"{ENDPOINT}/{dp.id}", json=payload)
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.updated"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_deleted_fires_event(self, mock_webhook, client):
+        mock_webhook.return_value = AsyncMock()
+        dp = DataProductFactory()
+
+        with webhook_v2_config():
+            response = client.delete(f"{ENDPOINT}/{dp.id}")
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.deleted"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_about_updated_fires_event(self, mock_webhook, client):
+        mock_webhook.return_value = AsyncMock()
+        dp = DataProductFactory()
+
+        with webhook_v2_config():
+            response = client.put(
+                f"{ENDPOINT}/{dp.id}/about", json={"about": "new text"}
+            )
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.about_updated"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_status_updated_fires_event(self, mock_webhook, client):
+        mock_webhook.return_value = AsyncMock()
+        dp = DataProductFactory()
+
+        with webhook_v2_config():
+            response = client.put(
+                f"{ENDPOINT}/{dp.id}/status", json={"status": "active"}
+            )
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.status_updated"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_setting_changed_fires_event(self, mock_webhook, client):
+        mock_webhook.return_value = AsyncMock()
+        dp = DataProductFactory()
+        setting = DataProductSettingFactory()
+
+        with webhook_v2_config():
+            response = client.post(
+                f"{ENDPOINT}/{dp.id}/settings/{setting.id}",
+                params={"value": "test-value"},
+            )
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.setting_changed"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_input_port_linked_fires_event(self, mock_webhook, client):
+        mock_webhook.return_value = AsyncMock()
+        dp = DataProductFactory()
+        output_port = DatasetFactory()
+
+        with webhook_v2_config():
+            response = client.post(
+                f"{ENDPOINT}/{dp.id}/link_input_ports",
+                json={"input_ports": [str(output_port.id)], "justification": "test"},
+            )
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.input_port_linked"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_input_port_unlinked_fires_event(self, mock_webhook, client):
+        mock_webhook.return_value = AsyncMock()
+        dp = DataProductFactory()
+        output_port = DatasetFactory()
+        DataProductDatasetAssociationFactory(data_product=dp, dataset=output_port)
+
+        with webhook_v2_config():
+            response = client.delete(f"{ENDPOINT}/{dp.id}/input_ports/{output_port.id}")
+
+        assert response.status_code == 200
+        mock_webhook.assert_awaited_once()
+        event_type, data = mock_webhook.call_args.args
+        assert event_type == "data_product.input_port_unlinked"
+        assert "data_product" in data
+
+    @pytest.mark.usefixtures("admin")
+    @pytest.mark.parametrize(
+        ("method", "path", "body"),
+        [
+            ("post", "/api/v2/data_products", {}),
+            ("put", "/api/v2/data_products/00000000-0000-0000-0000-000000000000", {}),
+            (
+                "delete",
+                "/api/v2/data_products/00000000-0000-0000-0000-000000000000",
+                None,
+            ),
+        ],
+    )
+    @patch("app.core.webhooks.v2.call_v2_webhook")
+    def test_no_event_fired_on_failure(self, mock_webhook, client, method, path, body):
+        mock_webhook.return_value = AsyncMock()
+        with webhook_v2_config():
+            kwargs = {"json": body} if body is not None else {}
+            getattr(client, method)(path, **kwargs)
+        mock_webhook.assert_not_awaited()
